fix: super admin general fixes
