+++ modules/super-admin/analytics-and-reporting/topSellingProduct.tsx
@@ -186,11 +186,11 @@ const TopSellingProducts: React.FC<TopSellingProps> = ({ dateRange, reportClicke
             </div>
             <div>
               <div className="border-b-[0.0625rem] border-[#EAECF0]"></div>
-              <SuperAdminPagination
+              {/* <SuperAdminPagination
                 currentPage={currentPage}
                 totalPages={Math.ceil(topSellingProducts?.length / pageSize)}
                 onPageChange={handlePageChange}
-              />
+              /> */}
               <div className="h-[0.94rem] rounded-b-[0.5rem] border-[0.001rem] border-t-0 border-[#EAECF0] max-[760px]:rounded-br-none"></div>
             </div>
           </div>
