+++ modules/super-admin/analytics-components/AnalyticsAndReportingTopSelling.tsx
@@ -80,13 +80,13 @@ export default function Page() {
               ))
             : 'Loading...'}
         </div>
-        <SuperAdminPagination
+        {/* <SuperAdminPagination
           currentPage={currentPage}
           totalPages={totalPages}
           onPageChange={(value) => {
             setCurrentPage(value);
           }}
-        />
+        /> */}
       </div>
     </section>
   );
