+++ modules/super-admin/components/vendormanagement/VendorStat.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { BannedDeletedVendorsProps } from '../../../../@types';
 import { LoadingText } from '../product-listing/ProductListingNavbar';
+import { formatNumber } from '../product-listing/product-details';
 const VendorsStat = ({
   showBanned,
   setShowBanned,
@@ -17,7 +18,7 @@ const VendorsStat = ({
             <p className="text-lg">Total Vendors</p>
           </div>
           <div className="flex items-center justify-between">
-            {isLoading ? <LoadingText /> : <h2 className="text-4xl font-bold">{data?.data?.length}</h2>}
+            {isLoading ? <LoadingText /> : <h2 className="text-4xl font-bold">{formatNumber(data?.total_shops) ?? 0}</h2>}
           </div>
         </div>
         <div className=" p-4 border-solid rounded-md border-white-115 border-2">
@@ -29,7 +30,7 @@ const VendorsStat = ({
               <LoadingText />
             ) : (
               <h2 className="text-4xl font-bold ">
-                {data?.data?.filter((item: any) => item?.vendor_status === 'Banned')?.length}
+                {formatNumber(data?.total_banned_shops) ?? 0}
               </h2>
             )}
             <button
@@ -45,7 +46,7 @@ const VendorsStat = ({
             <p className="text-lg text-gray-500">Deleted Vendors</p>
           </div>
           <div className="flex items-center justify-between">
-            {isLoading ? <LoadingText /> : <h2 className="text-4xl font-bold">{data?.data?.filter((item: any) => item?.vendor_status === 'Deleted')?.length}</h2>}
+            {isLoading ? <LoadingText /> : <h2 className="text-4xl font-bold">{formatNumber(data?.total_deleted_shops) ?? 0}</h2>}
             <button
               className="px-3 py-1 bg-brand-green-primary hover:bg-brand-green-hover text-white-100 rounded-2xl"
               onClick={() => setShowDeleted(!showDeleted)}
