+++ pages/super-admin/feedback-and-customer-support/general-complaints.tsx
@@ -641,7 +641,7 @@ function GeneralComplaints({ complain }: { complain: Complain }) {
             </div>
           </div>
 
-          <SuperAdminPagination currentPage={pageNum} totalPages={pageCount} onPageChange={handlePageClick} />
+          {/* <SuperAdminPagination currentPage={pageNum} totalPages={pageCount} onPageChange={handlePageClick} /> */}
         </div>
       </div>
     </>
