+++ pages/super-admin/product-listing/deleted-products/[id].tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import SuperAdminProdDetails from '@modules/super-admin/components/product-listing/product-details';
 import DeleteModal from '@modules/super-admin/components/product-listing/product-details/DeleteModal';
-import { useGetProdDetails } from '../../../../http';
+import { useGetProdDetails } from '../../../../http/super-admin1';
 import { useRouter } from 'next/router';
 import Loader from '@modules/portfolio/component/landing/Loader';
 import { toast } from 'react-toastify';
