+++ pages/super-admin/product-listing/deleted-products/index.tsx
@@ -4,7 +4,7 @@ import SearchProduct from '@modules/super-admin/components/product-listing/searc
 import { useEffect, useState } from 'react';
 import SuperAdminPagination from '@modules/super-admin/components/pagination';
 import { useRouter } from 'next/router';
-import { useGetProd } from '../../../../http';
+import { useGetProd } from '../../../../http/super-admin1';
 import { DeletedProducts } from '../../../../@types';
 import { LoadingTable } from '@modules/super-admin/components/product-listing/ProductListingTable';
 import { formatDate } from '@modules/super-admin/components/product-listing/product-details';
@@ -147,11 +147,11 @@ const SanctionedProducts = () => {
                     ))}
                   </tbody>
                 </table>
-                <SuperAdminPagination
+                {/* <SuperAdminPagination
                   currentPage={currentPage}
                   totalPages={totalPages}
                   onPageChange={handlePageChange}
-                />
+                /> */}
               </>
             ) : (
               <p className="text-red-100 my-10 w-fit mx-auto">Nothing to show</p>
