+++ pages/super-admin/vendor-management/index.tsx
@@ -8,12 +8,14 @@ import SuperAdminPagination from '@modules/super-admin/components/pagination';
 import SearchProduct from '@modules/super-admin/components/vendormanagement/SearchProduct';
 import FilterProduct from '@modules/super-admin/components/vendormanagement/FilterProduct';
 import Button from '@ui/Button';
-import { useGetAllVendor } from '../../../http';
 import { LoadingTable } from '@modules/super-admin/components/product-listing/ProductListingTable';
+import { useGetAllVendor } from '../../../http/super-admin1';
+
 const Index = () => {
+  const [currentPage, setCurrentPage] = useState(1);
+  // page: number, search: string, status: string
   const { data, isLoading } = useGetAllVendor();
   //Variables for the pagination
-  const [currentPage, setCurrentPage] = useState(1);
   const [showBanned, setShowBanned] = useState(false);
   const [showDeleted, setShowDeleted] = useState(false);
   const [searchVal, setSearchVal] = useState('');
@@ -139,11 +141,11 @@ const Index = () => {
                       ? deletedVendors?.map((data: any) => <VendorLists key={data?.id} data={data} />)
                       : visibleVendors?.map((data: any) => <VendorLists key={data?.id} data={data} />)}
                   </div>
-                  <SuperAdminPagination
+                  {/* <SuperAdminPagination
                     currentPage={currentPage}
-                    totalPages={totalPages}
-                    onPageChange={handlePageChange}
-                  />
+                    totalPages={data?.total_pages}
+                    setCurrentPage={setCurrentPage}
+                  /> */}
                 </>
               ) : (
                 <p className="text-red-100 my-10 w-fit mx-auto">Nothing to show</p>
